moved away the main loop to ceefax
