+++ page/Page.py
@@ -28,6 +28,9 @@ class Page(object):
         self.loaded = False
         self.title = ""
 
+        self.main_loop = None
+        self.keyboard_handler = None
+
     def now(self):
         return now.now()
